pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "Simple TODO web api to store your tasks"
authors = ["Lampower <123113309+Lampower@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standart"], version = "^0.115.5"}
uvicorn = {extras = ["standard"], version = "^0.32.1"}
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.36"
punq = "^0.7.0"
bcrypt = "^4.2.1"
pyjwt = "^2.10.1"

[tool.poetry.scripts]
start = "app.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
